ADD UptadeHiringDTO, UptadeStateHiringDTO and CRUD in the folders state and hiring

+++ api/src/hiring/hiring.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Hiring } from './hiring.entity';
 import { Repository } from 'typeorm';
 import { UserService } from 'src/user/user.service';
 import { StateHiringService } from './state/stateHiring.service';
+import { ResponseMessage } from 'src/utils/types/message.type';
 
 @Injectable()
 export class HiringService {
@@ -52,22 +53,28 @@ export class HiringService {
     }
   }
 
-  async deleteHire(id: number): Promise<void> {
+  async deleteHire(id: number): Promise<ResponseMessage> {
     try {
+      const hire = await this.hiringRepository.findOneBy({ id })
+      if (!hire) throw new NotFoundException ('El contrato no existe')
       await this.hiringRepository.delete(id);
+      return { message: 'El contrato se ha borrado correctamente' };
     } catch (error) {
-      console.error('Error al eliminar el contrato:', error);
+      console.error('Error al eliminar el contrato', error);
       throw new Error('Error al eliminar el contrato');
     }
   }
 
-  async updateHire(id: number, updatedData: Partial<Hiring>): Promise<Hiring> {
+  async updateHire(id: number, UpdateHireDTO): Promise<Hiring> {
     try {
-      await this.hiringRepository.update(id, updatedData);
+      const hireFound = this.hiringRepository.findOneBy({ id })
+      if (!hireFound) throw new NotFoundException ('El contrato no existe')
+      await this.hiringRepository.update(id, UpdateHireDTO);
       return await this.hiringRepository.findOneByOrFail({ id });
     } catch (error) {
       console.error('Error al actualizar el contrato:', error);
       throw new Error('Error al actualizar el contrato');
     }
+  
   }
 }
